Turn rust examples into packages
