+++ Cargo.toml
@@ -1,6 +1,11 @@
 [workspace]
 resolver = "2"
-members = ["rust/*", "python/foxglove-sdk"]
+members = [
+    "rust/foxglove",
+    "rust/foxglove-proto-gen",
+    "rust/examples/*",
+    "python/foxglove-sdk",
+]
 
 [workspace.dependencies]
 bytes = "1.9.0"
