+++ rust/examples/client-publish/src/main.rs
