+++ rust/examples/param-server/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "foxglove-param-server"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+foxglove = { path = "../../foxglove", features = ["unstable"] }
+tokio = { version = "1.0", features = ["full"] }
+tokio-util = { version = "0.7", features = ["rt"] }
+clap = { version = "4.5", features = ["derive"] }
+env_logger = "0.11.5"
