+++ rust/examples/ws-demo-blocking/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "foxglove-ws-demo-blocking"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+foxglove = { path = "../../foxglove" }
+clap = { version = "4.5", features = ["derive"] }
+env_logger = "0.11.5"
+ctrlc = "3.4.5"
+schemars = "0.8.21"
+serde = { version = "1.0", features = ["derive"] }
