+++ rust/examples/ws-demo/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "foxglove-ws-demo"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+foxglove = { path = "../../foxglove" }
+tokio = { version = "1.0", features = ["full"] }
+schemars = "0.8.21"
+serde = { version = "1.0", features = ["derive"] }
+clap = { version = "4.5", features = ["derive"] }
+env_logger = "0.11.5"
+prost-types = "0.13.4"
+quaternion = "2.0.0"
