+++ rust/examples/ws-demo/src/main.rs
@@ -1,22 +1,12 @@
+use clap::Parser;
+
 use foxglove::schemas::{
     Color, CubePrimitive, FrameTransform, Pose, Quaternion, SceneEntity, SceneUpdate, Vector3,
 };
 use foxglove::static_typed_channel;
 use schemars::JsonSchema;
 use serde::Serialize;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
 use std::time::Duration;
-
-fn euler_to_quaternion(roll: f64, pitch: f64, yaw: f64) -> Quaternion {
-    let quat = quaternion::euler_angles(roll, pitch, yaw);
-    Quaternion {
-        x: quat.1[0],
-        y: quat.1[1],
-        z: quat.1[2],
-        w: quat.0,
-    }
-}
 
 #[derive(Debug, Serialize, JsonSchema)]
 struct Message {
@@ -27,17 +17,6 @@ struct Message {
 static_typed_channel!(pub BOX_CHANNEL, "/boxes", SceneUpdate);
 static_typed_channel!(pub TF_CHANNEL, "/tf", FrameTransform);
 static_typed_channel!(pub MSG_CHANNEL, "/msg", Message);
-
-#[allow(dead_code)]
-pub fn log_blocking(fps: u8, stop: Arc<AtomicBool>) {
-    let mut counter: u32 = 0;
-    let duration = Duration::from_millis(1000 / u64::from(fps));
-    while !stop.load(Ordering::Relaxed) {
-        log(counter);
-        std::thread::sleep(duration);
-        counter += 1;
-    }
-}
 
 #[allow(dead_code)]
 pub async fn log_forever(fps: u8) {
@@ -96,4 +75,46 @@ pub fn log(counter: u32) {
         rotation: Some(euler_to_quaternion(1.0, 0.0, f64::from(counter) * 0.1)),
         ..Default::default()
     });
+}
+
+#[derive(Debug, Parser)]
+struct Cli {
+    /// Server TCP port.
+    #[arg(short, long, default_value_t = 8765)]
+    port: u16,
+    /// Server IP address.
+    #[arg(long, default_value = "127.0.0.1")]
+    host: String,
+    /// Frames per second.
+    #[arg(long, default_value_t = 60)]
+    fps: u8,
+}
+
+#[tokio::main]
+async fn main() {
+    let env = env_logger::Env::default().default_filter_or("debug");
+    env_logger::init_from_env(env);
+
+    let args = Cli::parse();
+
+    let server = foxglove::WebSocketServer::new()
+        .name("ws-demo")
+        .bind(&args.host, args.port)
+        .start()
+        .await
+        .expect("Server failed to start");
+
+    tokio::task::spawn(log_forever(args.fps));
+    tokio::signal::ctrl_c().await.ok();
+    server.stop().await;
+}
+
+fn euler_to_quaternion(roll: f64, pitch: f64, yaw: f64) -> Quaternion {
+    let quat = quaternion::euler_angles(roll, pitch, yaw);
+    Quaternion {
+        x: quat.1[0],
+        y: quat.1[1],
+        z: quat.1[2],
+        w: quat.0,
+    }
 }
