+++ rust/foxglove/Cargo.toml
@@ -35,19 +35,7 @@ tracing.workspace = true
 [dev-dependencies]
 assert_matches = "1.5.0"
 clap = { version = "4.5", features = ["derive"] }
-ctrlc = "3.4.5"
 env_logger = "0.11.5"
 futures-util = "0.3.31"
-quaternion = "2.0.0"
 tempfile = "3.15.0"
 tracing-test = "0.2.5"
-
-[[example]]
-name = "client-publish"
-path = "examples/unstable/client-publish.rs"
-required-features = ["unstable"]
-
-[[example]]
-name = "param-server"
-path = "examples/unstable/param-server.rs"
-required-features = ["unstable"]
